requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
Pillow>=10.3.0
mss>=9.0.1
pytest>=8.2.0
pyautogui>=0.9.54
playwright>=1.49.0
PySide6>=6.7.2
